Moved kubeconfig loading logic to testutils
