+++ e2e/src/test/java/io/kubernetes/client/e2e/ssl/SSLTest.java
@@ -12,6 +12,7 @@ limitations under the License.
 */
 package io.kubernetes.client.e2e.ssl;
 
+import io.kubernetes.client.e2e.util.TestUtils;
 import io.kubernetes.client.informer.SharedIndexInformer;
 import io.kubernetes.client.informer.SharedInformerFactory;
 import io.kubernetes.client.informer.cache.Lister;
@@ -25,18 +26,10 @@ import nl.altindag.log.LogCaptor;
 import nl.altindag.log.model.LogEvent;
 import org.junit.jupiter.api.Test;
 
-import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
 import java.util.Base64;
 import java.util.Optional;
 
-import static io.kubernetes.client.util.KubeConfig.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
@@ -74,7 +67,7 @@ class SSLTest {
 
             logCaptor.clearLogs();
 
-            Optional<byte[]> certificateAuthorityData = findConfigInHomeDir().flatMap(SSLTest::getCertificateAuthorityData);
+            Optional<byte[]> certificateAuthorityData = getCertificateAuthorityDataFromKubeConfig();
             assertThat(certificateAuthorityData).isPresent();
             apiClient.setSslCaCert(new ByteArrayInputStream(certificateAuthorityData.get()));
 
@@ -92,17 +85,10 @@ class SSLTest {
         }
     }
 
-    private static Optional<byte[]> getCertificateAuthorityData(File configInHomeDir) {
-        try (InputStream inputStream = Files.newInputStream(configInHomeDir.toPath());
-             InputStreamReader inputStreamReader = new InputStreamReader(inputStream, StandardCharsets.UTF_8);
-             BufferedReader bufferedReader = new BufferedReader(inputStreamReader)) {
-
-            KubeConfig kubeConfig = KubeConfig.loadKubeConfig(bufferedReader);
-            String certificateAuthorityData = kubeConfig.getCertificateAuthorityData();
-            return Optional.of(Base64.getDecoder().decode(certificateAuthorityData));
-        } catch (IOException e) {
-            return Optional.empty();
-        }
+    private static Optional<byte[]> getCertificateAuthorityDataFromKubeConfig() {
+        return TestUtils.getKubeConfig()
+                .map(KubeConfig::getCertificateAuthorityData)
+                .map(base64EncodedCertificateAuthority -> Base64.getDecoder().decode(base64EncodedCertificateAuthority));
     }
 
     private static Optional<Throwable> getCertificateNotFoundException(LogCaptor logCaptor) {
@@ -112,48 +98,6 @@ class SSLTest {
                 .map(Optional::get)
                 .filter(exception -> exception.getMessage().contains("unable to find valid certification path to requested target"))
                 .findAny();
-    }
-
-    private static Optional<File> findConfigInHomeDir() {
-        final File homeDir = findHomeDir();
-        if (homeDir != null) {
-            final File config = new File(new File(homeDir, KUBEDIR), KUBECONFIG);
-            if (config.exists()) {
-                return Optional.of(config);
-            }
-        }
-        return Optional.empty();
-    }
-
-    private static File findHomeDir() {
-        final String envHome = System.getenv(ENV_HOME);
-        if (envHome != null && envHome.length() > 0) {
-            final File config = new File(envHome);
-            if (config.exists()) {
-                return config;
-            }
-        }
-        if (System.getProperty("os.name").toLowerCase().startsWith("windows")) {
-            String homeDrive = System.getenv("HOMEDRIVE");
-            String homePath = System.getenv("HOMEPATH");
-            if (homeDrive != null
-                    && homeDrive.length() > 0
-                    && homePath != null
-                    && homePath.length() > 0) {
-                File homeDir = new File(new File(homeDrive), homePath);
-                if (homeDir.exists()) {
-                    return homeDir;
-                }
-            }
-            String userProfile = System.getenv("USERPROFILE");
-            if (userProfile != null && userProfile.length() > 0) {
-                File profileDir = new File(userProfile);
-                if (profileDir.exists()) {
-                    return profileDir;
-                }
-            }
-        }
-        return null;
     }
 
 }
